Renamed .jsx files to .js, additions to ImageControl component
