+++ client/app/components/WikiEditor/Toolbar/InlineControls.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ToolbarButton from './ToolbarButton.jsx';
+import ToolbarButton from './ToolbarButton.js';
 
 const INLINE_TYPES = [
 	{label: 'Bold', fa: 'fa-bold', style: 'BOLD'},
