+++ client/app/components/WikiEditor/entities/Image.js
@@ -1,0 +1,8 @@
+const Image = (props) => {
+  const {src, alt} = props.contentState.getEntity(props.entityKey).getData();
+  return (
+    <img src={src} alt={alt} />
+  );
+}
+
+export default Image;
