+++ client/app/components/WikiEditor/entities/index.js
@@ -1,0 +1,4 @@
+import Link from './Link';
+import Image from './Image';
+
+export {Image, Link};
