+++ client/app/components/WikiEditor/helpers/getEntities.js
@@ -1,15 +1,15 @@
-const getLinkEntities = (contentBlock, callback, contentState) => {
+const getEntities = (entityType) => (contentBlock, callback, contentState) => {
   contentBlock.findEntityRanges(
     (character) => {
       const entityKey = character.getEntity();
 
       return (
         entityKey !== null &&
-        contentState.getEntity(entityKey).getType() === 'LINK'
+        contentState.getEntity(entityKey).getType() === entityType
       );
     },
     callback
   );
 }
 
-export default getLinkEntities;
+export default getEntities;
